+++ internal/storemigrate/detector_darwin.go
@@ -1,0 +1,11 @@
+//go:build darwin
+
+package storemigrate
+
+import "io/fs"
+
+type platformCloudPlaceholderDetector struct{}
+
+func (platformCloudPlaceholderDetector) IsCloudOnly(_ string, info fs.FileInfo) bool {
+	return fileInfoDataless(info)
+}